refactor: move service + hooks + types to infradir
